requirements.txt
supabase==1.0.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9
selenium==4.15.0
requests==2.31.0 